+++ packages/components/dummy/src/Dummy.tsx
@@ -1,10 +1,14 @@
-import { ComponentPropsWithoutRef, PropsWithChildren } from 'react'
+import { PropsWithChildren } from 'react'
 
-export interface DummyProps extends ComponentPropsWithoutRef<'span'> {
+export interface DummyProps {
   /* type of dummy */
   type: 'bar' | 'foo'
 }
 
 export function Dummy({ type = 'bar', ...rest }: PropsWithChildren<DummyProps>) {
-  return <span {...rest}>{type}</span>
+  return (
+    <span {...rest} className="text-body-1">
+      {type}
+    </span>
+  )
 }