+++ packages/utils/theme/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.4](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils@2.6.3...@spark-ui/theme-utils@2.6.4) (2023-03-07)
+
+**Note:** Version bump only for package @spark-ui/theme-utils
+
+## [2.6.3](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils@2.6.2...@spark-ui/theme-utils@2.6.3) (2023-03-07)
+
+**Note:** Version bump only for package @spark-ui/theme-utils
+
+## [2.6.2](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils@2.6.1...@spark-ui/theme-utils@2.6.2) (2023-03-03)
+
+**Note:** Version bump only for package @spark-ui/theme-utils
+
+## [2.6.1](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils@2.6.0...@spark-ui/theme-utils@2.6.1) (2023-03-03)
+
+**Note:** Version bump only for package @spark-ui/theme-utils
+
 # [2.6.0](https://github.com/adevinta/spark/compare/@spark-ui/theme-utils@2.5.1...@spark-ui/theme-utils@2.6.0) (2023-03-02)
 
 ### Features